+++ src/system/FontSDL.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 
 #include "global.h"
-#include "FontSDL.h"
+#include "system/FontSDL.h"
 
 bool FontSDL::mFontInit = false;
 TTF_Font* FontSDL::mFont = nullptr;